Journal app
In this journal I just want to make a dinamic app for every user, with a register form and a login form with a function to send firebase the user unique id and the rest of the data. 
As well I worked with hooks like useState, useForm, useMemo, useCallback, and others. Also Redux, Sass for the styles and Postman.